initial restructuring of image pyramid
